requirements.txt
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.61.1
fpdf2==2.8.5
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
pillow==12.1.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.2 # Added for Telegram notifications
six==1.17.0
